api: fix pxe_reboot support; add {disk,bios}_reboot

+++ mr_provisioner/models.py
@@ -309,11 +309,27 @@ class Machine(db.Model):
         else:
             bmc.type_inst.set_power(self, "on")
 
-    def set_power(self, power_state):
-        power_state = 'reset' if power_state == 'reboot' else power_state
+    def bios_reboot(self):
         bmc = self.bmc
         # XXX: raise exception if not bmc
-        bmc.type_inst.set_power(self, power_state)
+        bmc.type_inst.set_bootdev(self, "bios")
+        if self.power_state == "on":
+            bmc.type_inst.set_power(self, "reset")
+        else:
+            bmc.type_inst.set_power(self, "on")
+
+    def set_power(self, power_state):
+        if power_state == 'pxe_reboot':
+            self.pxe_reboot()
+        elif power_state == 'disk_reboot':
+            self.disk_reboot()
+        elif power_state == 'bios_reboot':
+            self.bios_reboot()
+        else:
+            power_state = 'reset' if power_state == 'reboot' else power_state
+            bmc = self.bmc
+            # XXX: raise exception if not bmc
+            bmc.type_inst.set_power(self, power_state)
 
     def deactivate_sol(self):
         bmc = self.bmc
